added home image
